backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.2
pytest==8.3.3
pytest-asyncio==0.24.0
slowapi==0.1.9
limits==3.13.0
